requirements.txt
azure-functions==1.17.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-identity==1.15.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
httpx==0.25.2
freezegun==1.2.2
factory-boy==3.3.0 